hafta 18

+++ Core/Utilities/Helper/FileHelper/FileHelper.cs
@@ -2,21 +2,25 @@
 using Core.Utilities.Helper.GuidHelper;
 using Microsoft.AspNetCore.Http;
 using System;
+using System.IO;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Core.Constants;
 
 namespace Core.Utilities.Helper.FileHelper
 {
     public class FileHelper : IFileHelper
     {
+    
+
         public string Add(IFormFile file)
         {
             //dosyanın uzantısını alıyorum.
             string fileExtension = Path.GetExtension(file.FileName);
             //Guid ile uzantıyı birleştiriyorum.
-            string uniqueFileName = GuidHelper.GuidHelper.Create() + fileExtension;
+            string uniqueFileName = GuidHelper_.Create() + fileExtension;
             //GuidHelper_.Create() + fileExtension;
             //kaydetmek istediğim yerin tam yolunu alıyorum.
             var imagePath = FilePath.Full(uniqueFileName);
@@ -32,20 +36,20 @@ namespace Core.Utilities.Helper.FileHelper
 
         public void Delete(string path)
         {
-           /* if (Path.Exists(FilePath.Full(path))) //varsa true
+            if (File.Exists(FilePath.Full(path))) //varsa true
             {
                 File.Delete(FilePath.Full(path));
             }
             else
             {
-                throw new DirectoryNotFoundException();
-            }*/
+                throw new DirectoryNotFoundException(Messages.FileNotFound);
+            }
         }
 
         public void Update(IFormFile file, string imagePath)
         {
-            /*var fullpath = FilePath.Full(imagePath);
-            if (Path.Exists(fullpath))
+            var fullpath = FilePath.Full(imagePath);
+            if (File.Exists(fullpath))
             {
                 using FileStream fileStream = new(fullpath, FileMode.Create);
                 file.CopyTo(fileStream);
@@ -54,7 +58,7 @@ namespace Core.Utilities.Helper.FileHelper
             else
             {
                 throw new DirectoryNotFoundException(Messages.FileNotFound);
-            }*/
+            }
         }
     }
 }
